Move SchemaRequestDetails in shared components directory.
