+++ coral/src/app/features/components/SchemaRequestDetails.tsx
